\name{powell}
\alias{powell}
\alias{Powell}
\alias{powell_counts}
\docType{data}
\title{
Dataset due to Powell (1990)
}
\description{
Dataset due to Powell (1990)
}
\usage{data(powell)}
\format{
A frequency table of counts of association data.
}
\source{
  \itemize{
    \item W. Powell, M. Coleman and J. McNicol 1990  \dQuote{The
    statistical analysis of potato culture data}.
      \emph{Plant Cell, Tissue and Organ Culture} 23:159-164
      }
}
\examples{
data(powell)
Lindsey(powell, powell_counts)
}
\keyword{datasets}
